requirements.txt
# ========================================
# AI Teaching Assistant - Dependencies
# ========================================

# Core LangChain and LangGraph
langgraph>=0.2.0
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-community>=0.3.0
langchain-chroma>=0.1.0

# Vector Store
chromadb>=0.4.0

# Data validation
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Environment management
python-dotenv>=1.0.0

# Utilities
python-dateutil>=2.8.0

# ========================================
# API Dependencies
# ========================================

# FastAPI & Web Server
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
websockets>=12.0
python-multipart>=0.0.6

# ========================================
# Voice Agent Dependencies (LiveKit)
# ========================================

# Core LiveKit Agents
livekit-agents
livekit-plugins-openai
livekit-plugins-silero
livekit-plugins-langchain
livekit-plugins-deepgram
livekit-plugins-turn-detector
livekit-plugins-cartesia

# Async support
aiohttp
